+++ protocol/CopyObjectRequest.go
@@ -1,0 +1,7 @@
+package protocol
+
+// CopyObjectRequest is a subset of Object for use to disallow providing certain fields.
+type CopyObjectRequest struct {
+	// ID is the unique identifier for this object in Object Drive.
+	ID string `json:"id"`
+}